pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ryzen-vf-analysis"
version = "0.1.0"
description = "Reusable Ryzen 9700X V/F analysis package and notebook workflow."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "polars>=1.38.1",
    "pylint>=3.3.8",
    "pyarrow>=23.0.1",
    "pyright>=1.1.408",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["vfanalysis*", "vf_analysis*"]

[tool.ruff]
target-version = "py314"
line-length = 100
extend-exclude = ["notebooks/vf_exploration.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "C4"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pylint.main]
py-version = "3.14"
ignore-paths = ["^src/ryzen_vf_analysis\\.egg-info/.*"]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "too-many-arguments",
    "too-many-locals",
]
